Fix keyboard/button lights with AutoBrightness
Keyboard activity must always trigger key/button events. The point is
that they should use the autobrightness values in those cases. And the
decision should be based on whether the phone is currently in auto
mode, not whether or not the phone *supports* auto mode. More weak-minded
Google coders at work...